% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}

\item{W}{is the total size of the knapsack}
}
\description{
function knapsack_brute_force(x, W) that takes a data.frame cx with two variables v and w and returns the maximum knapsack value and which elements
}
\details{
Brute force algorithm for the knapsack problem

This algorithm gives all possbile values with good accuracy , and also gives the maximum value for the knapsack
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\seealso{
\code{\link{Brute Force Knapsack}}
}
\author{
Omid Lavakhamseh, Hamed Godazgar
}
