% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_heuristic.r
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Implements Greedy ALgorithm for knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{weight threshold for the knapsack}
}
\value{
Total v and position of w's for that value
}
\description{
Implements Greedy ALgorithm for knapsack problem
}
\seealso{
\code{\link{brute_force_knapsack}}
\code{\link{knapsack_dynamic}}
}
