% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic_programming.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Implements Dynamic programming for knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{Knapsack capacity}
}
\value{
Total and Maximum value within the max capacity
}
\description{
Implements Dynamic programming for knapsack problem
}
